% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadlock.R
\name{hadlock_efw}
\alias{hadlock_efw}
\title{Estimate fetal weight with Hadlock's three-parameter formula}
\usage{
hadlock_efw(headcirc_cm, abdocirc_cm, femurlen_cm)
}
\arguments{
\item{headcirc_cm, abdocirc_cm, femurlen_cm}{Numeric vectors of length one or
more with head circumference/abdominal circumference/femur length
measurements in centimetres. These inputs should be recyclable by
`vctrs::vec_recycle_common()`.}
}
\value{
A numeric vector the same length as the inputs, with Hadlock's
  estimated fetal weight values in grams.
}
\description{
Estimate fetal weight with Hadlock's three-parameter formula
}
\examples{
hadlock_efw(5, 5, 5)
}
\references{
Hadlock F, Harrist R, Sharman R, et al. **Estimation of fetal weight with the
use of head, body, and femur measurements — a prospective study.** *Am J
Obstet Gynecol.* 1985, **151(3):333–7)**.
}
