% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{value2centile}
\alias{value2centile}
\alias{value2zscore}
\title{Convert values to z-scores/centiles in the WHO Fetal Standards}
\usage{
value2centile(y, gest_days, sex = NULL, acronym)

value2zscore(y, gest_days, sex = NULL, acronym)
}
\arguments{
\item{y}{Numeric vectors of fetal biometry values. These should have the
units specified below for `acronym`.}

\item{gest_days}{A numeric vector of lenght one or more with gestational ages
in days. Elements should be between 98 and 280 days (14 to 40 weeks).}

\item{sex}{A character vector of "M" and "F" values (case-sensitive). Only
used for the WHO estimated fetal weight (EFW)-for-GA standard, which uses
sex-specific coefficients for each quantile where `sex` is either `"M"` or
`"F"`. An error will be thrown if there are values in `sex` which are not
`"M"` or `"F"`.}

\item{acronym}{A single string denoting the WHO Fetal Standard in use.
  Available acronyms are:
  * "acfga" - abdominal circumference (mm)-for-GA
  * "bpdfga" - biparietal diameter (mm)-for-GA
  * "efwfga" - estimated fetal weight (grams)-for-GA
  * "flfga" - femur length (mm)-for-GA
  * "hcfga" - head circumference (mm)-for-GA
  * "hlfga" - humerus length (mm)-for-GA

  `acronym` is case-sensitive, and an error will be thrown if `acronym` is
  not one of the above values.}
}
\value{
A numeric vector of centiles (`value2centile()`) or z-scores
  `value2zscore()`. Where centiles/z-scores would be very extreme (i.e.
  above the)
}
\description{
Convert values to z-scores/centiles in the WHO Fetal Standards
}
\examples{
# Get a centile for a femur length value
value2centile(30, 20 * 7, acronym = "flfga")

# Get a centile for an estimated fetal weight value in males
value2centile(330, 20 * 7, "M", "efwfga")

# Get the equivalent z-score
value2zscore(330, 20 * 7, "M", "efwfga")

# The function is vectorised and will recycle inputs
value2zscore(seq(300, 390, 30), 20 * 7, "F", "efwfga")
}
\references{
Kiserud T, Piaggio G, Carroli G, et al. **The World Health Organization fetal
growth charts: a multinational longitudinal study of ultrasound biometric
measurements and estimated fetal weight.** *PLoS Med* 2017, **14:e1002220**.
\doi{10.1371/journal.pmed.1002220}

Kiserud T, Benachi A, Hecher K, et al. **The World Health Organisation fetal
growth charts: concept, findings, interpretation, and application.** *Am J
Obstet Gynecol.* 2018, **218(2S):S619-29**. \doi{10.1016/j.ajog.2017.12.010}
}
