% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whoFetalStandards-package.R
\docType{package}
\name{whoFetalStandards-package}
\alias{whoFetalStandards-package}
\alias{whoFetalStandards}
\title{whoFetalStandards: WHO Standards for Fetal Growth}
\description{
Convert fetal biometry measures to z-scores/centiles in the WHO Fetal growth
standards.
}
\examples{
\dontrun{
help(package = "whoFetalStandards")
browseURL("https://www.github.com/simpar1471/whoFetalStandards")
}
}
\author{
\strong{Maintainer}: Simon Parker \email{simon.parker1471@outlook.com} (\href{https://orcid.org/0009-0003-8214-4496}{ORCID})

}
\keyword{internal}
